+++ src/components/CommentList/CommentList.tsx
@@ -7,7 +7,7 @@ import { COMMENTS_FETCH_COUNT } from '@/lib/constants';
 import { useReadComments } from '@/lib/hooks';
 import { type CommentWithRelationsAndRelationCountsAndUserReaction } from '@/lib/types';
 
-import { CommentCard } from '../CommentCard';
+import CommentCard from '../CommentCard/CommentCard';
 import CommentCardSkeleton from '../CommentCardSkeleton/CommentCardSkeleton';
 import usePostCard from '../PostCard/usePostCard';
