+++ src/lib/utilities/authenticateUser.ts
