+++ src/lib/utilities/index.ts
@@ -4,4 +4,5 @@ export { default as parsePayload } from './parsePayload';
 export { default as getCommentQueryKey } from './getCommentQueryKey';
 export { default as getPostQueryKey } from './getPostQueryKey';
 
+export { default as cn } from './cn';
 export { default as getName } from './getName';
